WorkStatusForNetTrafficWithDjango/requirements.txt
amqp==2.1.4
backports.shutil-get-terminal-size==1.0.0
bcrypt==3.1.0
billiard==3.5.0.2
celery==4.0.2
cffi==1.8.2
decorator==4.0.10
Django==1.10.1
django-redis-cache==1.7.0
django-redis-sessions==0.5.6
djangorestframework==3.4.7
ecdsa==0.13
enum34==1.1.6
Fabric==1.12.0
gevent==1.2.0
greenlet==0.4.11
gunicorn==19.6.0
ipython-genutils==0.1.0
jieba==0.38
kombu==4.0.2
meld3==1.0.2
MySQL-python==1.2.5
numpy==1.11.3
paramiko==1.17.2
pathlib2==2.1.0
pexpect==4.2.1
pickleshare==0.7.4
prompt-toolkit==1.0.7
ptyprocess==0.5.1
pycparser==2.14
pycrypto==2.6.1
Pygments==2.1.3
PyJWT==1.4.2
pytz==2016.10
redis==2.10.5
scikit-learn==0.18.1
scipy==0.18.1
simplegeneric==0.8.1
six==1.10.0
sklearn==0.0
supervisor==3.3.1
traitlets==4.3.1
vine==1.1.3
wcwidth==0.1.7
